+++ Client/shared/index.ts
@@ -23,4 +23,5 @@ export * from './http/http-gateway.service';
 export * from './rx/rx-context.directive';
 
 export * from './meta/meta.service';
+export * from './meta/meta.store';