+++ Client/typings.d.ts
@@ -40,7 +40,7 @@ interface SystemJS {
 // Extra variables that live on Global that will be replaced by webpack DefinePlugin
 declare var ENV: string;
 declare var HMR: boolean;
-declare var Zone: {current: any};
+// declare var Zone: {current: any};
 interface GlobalEnvironment {
   ENV;
   HMR;